progress: invoke onclose every time display is stopped
During the refactor, it was changed to only call `onclose` when the
printer itself was closed. The original version invoked this when the
display was stopped in anyway.

Changing this back to see if it fixes a display issue that's popped up
with the refactor.

Restores the metric writer which accidentally got left off in the
refactor and wasn't noticed.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>