bake: Skip bake context reuse if `BUILDKIT_CONTEXT_KEEP_GIT_DIR=1`
When a remote Git URL is used with `buildx bake`, `.git` directories are
always omitted (see `bake.ReadRemoteFiles`).

Targets that rely on the default bake context should still be able to
change this behavior by passing the `BUILDKIT_CONTEXT_KEEP_GIT_DIR=1`
build argument. Check for this argument and skip the use of the bake
context state if the value is set to true.

Signed-off-by: Dan Duvall <dduvall@wikimedia.org>