store: remove uses of docker namesgenerator
It will not be part of the api or client modules, so replace it with
a local implementation, that's not using names of hackers/scientists,
so that we don't have to carry the debates around those choices here.

Thanks to ChatGPT for generating some lists (I haven't fully checked
them).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>