driver/docker-container: remove conditional UsernsMode=host for userns
This special handling was added in 5f8600f09865ece328641f965e1e8af65dd37ba1,
to work around an issue in the daemon. That issue was fixed in [moby@a826ca3],
which was backported to docker v20.10.13 in [moby@660b996].

Given that docker 20.10 reached EOL, and any currently supported version
of docker would have the fix from [moby@a826ca3], we can remove this special
handling.

[moby@a826ca3]: https://github.com/moby/moby/commit/a826ca3aefbd4d29344d851723731e3809f2a4ad
[moby@660b996]: https://github.com/moby/moby/commit/660b9962e4d2d4ad68898b00571a43127b84f6c3

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>