Ensure typed variables with no value still carry type
A value-less, untyped variable has always been converted to an empty
string.  The intention was that value-less, typed variables convert to a
 typed null, which was even specified in a code comment, but was never
actually implemented.

This resulted in a null value with a nil type.  A value with a nil type
cannot be coerced ("unified") with any other standard types.  When this
mismatch occurs, HCL attempts to return a diagnostic error, which in
turn panics as the nil type is literally a nil pointer.

Signed-off-by: Roberto Villarreal <rrjjvv@yahoo.com>