dap: attempt to find multiple digests when a failure happens
The solve error that gets returned sometimes has a platform when the LLB
returned by `ToState` doesn't. In order to ensure we find the failed
digest, we now search for both the digest of what was returned and also
clear out the platform and search for the digest without that in case
that one matches instead.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>