docs: add a link from the cli reference to the bake file reference for variables
I spent some time looking how to override a variable in a bakefile. My
assumption was that there would be a CLI option to do that and was
confused when I couldn't find it. This adds a reference to point someone
in the correct direction of using environment variables which is present
in the bake file reference but isn't in the CLI documentation.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>